[STF] clang-tidy batch 1: the mechanical findings, by script
From the first tidy run over STF (1237 findings, 214 TUs), the categories
that a script can apply from the reported location:

  misc-const-correctness            690  const at the declaration
  bugprone-narrowing-conversions    280  explicit static_cast of the operand
                                         (whole right-hand side for integral
                                         targets; the integer leaf inside a
                                         floating expression otherwise)
  cert-err34-c                       92  atoi/atol/atoll/atof -> std::sto*
  clang-diagnostic-braced-scalar-init 25 {N} -> N
  modernize-deprecated-headers       10  <stdio.h> -> <cstdio> etc.

plus four narrowing sites with dependent target types by hand.

Also: the negative-compile TUs under test/stf/static_error_checks are
marked NO_CLANG_TIDY (they fail to compile on purpose), and the
bugprone-exception-escape findings in exception_policy.cuh are
suppressed with a stated reason: clang-tidy's device pass compiles with
exceptions disabled, which makes every function implicitly non-throwing,
so the unit tests that throw on purpose and the ON_THROW(abort) runner
in invoke_nothrow are flagged for the throw the policy exists to catch.

+++ cudax/test/stf/freeze/freeze_write_back.cu
@@ -23,7 +23,7 @@ int main()
   int array[1024];
   for (size_t i = 0; i < 1024; i++)
   {
-    array[i] = 2 - i * i;
+    array[i] = static_cast<int>(2 - i * i);
   }
 
   auto lA = ctx.logical_data(array).set_symbol("A");
